Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/worklets/paint-worklet-service-worker-interception.https-expected.txt

CONSOLE MESSAGE: Error: Script error.
CONSOLE MESSAGE: Error: Script error.
CONSOLE MESSAGE: Error: Script error.

PASS addModule() on a controlled document should be intercepted by a service worker.
FAIL addModule() on a non-controlled document should not be intercepted by a service worker even if the script is under the service worker scope. assert_equals: expected "REJECTED" but got "RESOLVED"
PASS Static import should be intercepted by a service worker.
